Print: add printf method
Holy heck how did this not already exist?

It uses vsnprintf to load a buffer and the writes that buffer, and that
buffer is declared static and never free'd, so it's not necessarily the
most efficient or "proper" implementation, but it works. If all Print
instances share the same buffer that's efficient, and we don't have to
care about thread-safety.

fopencookie might work, but it might be worse by adding extra backend
buffer management (plus extra code size). An embedded printf
implementation with customizable putc/puts functions would be great, but
it's easier to just use the libc.