bootloader reset: emulate a double-tap rather than erasing the reset handler address
The current implementation of initiateReset/banzai erases the
application's reset handler address from the exception vector table in
flash, killing it permanently (stuck in the bootloader forever until
reprogrammed again).

However, we can use the same RAM address and magic value of the
bootloader's double-tap logic to reset into the bootloader without
disabling the application.

Doing so permanently uses 4 bytes of RAM, but that's OK. The code
assumes that the linker script will provide __bootloader_trap_reg
pointing to that address, but if not the weak symbol will resolve to
NULL and the old behavior will be used instead.

Also, drop the long_call and .ramfunc section attributes from banzai,
the linker script doesn't do anything with .ramfunc and the code was
just going to .text with everything else anyway. AFAIK there's no need
for that function to be special and live in RAM rather than flash.