Makefile: support BOSSA v1.9
Newer BOSSA versions can trigger the 1200 baud bootloader reset trick
without needing a external script (with my patch to fix it by adding
a delay to wait for the reset to actually happen), and that's quicker
than the ard-reset-arduino script.

BOSSA v1.9 also stopped assuming a 0x2000 flash starting address, so
specifying --offset=0x2000 is required or else it will start erasing the
bootloader code and brick the board.