Add basic Neostrip library
template class with a static number of pixels.

Unfortunately, the SAMD21 SPI MOSI pin idles high, which makes things
tricky. Adafruit gets around this by constantly DMA-ing out to the SPI
bus, which is clever and I'd probably grab that library if I ever really
wanted to use NeoPixels on the SAMD21.

I took the hardware route and am just running the data line through an
inverter and inverting all the out bits. 74LS04 has tens of MHz of
bandwidth, which is plenty for this.