bootloader: build with APP=1 to run the SAM-BA monitor as an application
Move the flash start address to 0x2000 and disable
check_start_application() to allow running the SAM-BA monitor as
a normal application. This configuration could be used to reflash the
true bootloader at address 0 without needing an external programmer.

To avoid duplicating the linker script, set the flash start address and
length from the Makefile and pass to the linker via --defsym.

To help distinguish this from the true 0x0000 bootloader, turn on the
green tx LED instead of the blue status LED in app mode.