adc test: set all the registers myself, add some averaging
I could software reset the ADC, but then might need to re-program the
calibration registers, so I'll leave it on. I hard-set all the registers
that matter anyway.

The 16-bit oversampling+decimation is cool, but requires 256 samples for
the full 16 bits, which takes a very long time since the prescaler and
sample time are maxed out. Multiple decimated shorter samples were also
more inaccurate than a single longer sample.

As a compromise, this commit averages 8 full-length samples, which takes
3.3ms and is pretty stable and accurate.