bbqshield: rewrite with fading logic; comments and cleanup
The problem with randomly fluctuating each pixel is that they tend to
stay roughly the same color over time (since the random drift is
uniformly centered around zero)

Instead, randomize the whole strip twice to get a starting and ending
color for each pixel and fade smoothly between those two endpoints.

With this we have ~100Hz updates of the strip and a single fade takes
about 2.5 seconds to go through all 257 steps. There's definitely some
color compression going on between the fading and the cie1931
correction, but it's not noticeable. Also not noticeable is the extra
250us it takes between fades to come up with the next fade endpoints.
I could hook up a timer for precise timing, but that's too much effort
to be worth it and the tiny irregularity can't be seen.