Makefile: explicitly link key system objects
cortex_handlers.c, main.cpp, startup.c, and variant.cpp are all code
that should always be included, and I was never sure how it was working
in the first place. I guess the linker script pulled in enough symbols
that recursively all the needed code got included.

This change has no impact on the binary size, but it just feels better.