bbqshield: move animations to GradientAnimation
It works and looks basically the same! Brightness fading is still
stacked on top in the main loop(), but the gradient color fading is now
in GradientAnimation.h.

Because of templates, everything has to be in headers, oh well.

I haven't benchmarked how much the class pointers and vtables slow this
down, but visually I can't really tell a difference.